Feat : 친구 기능 추가 및 수정

+++ dondoc/backend/src/main/java/com/dondoc/backend/user/service/FriendServiceImpl.java
@@ -3,6 +3,7 @@ package com.dondoc.backend.user.service;
 import com.dondoc.backend.common.exception.NotFoundException;
 import com.dondoc.backend.user.dto.friend.FriendListDto;
 import com.dondoc.backend.user.dto.friend.FriendRequestDto;
+import com.dondoc.backend.user.dto.friend.FriendSearchDto;
 import com.dondoc.backend.user.entity.Friend;
 import com.dondoc.backend.user.entity.User;
 import com.dondoc.backend.user.repository.FriendRepository;
@@ -10,6 +11,7 @@ import com.dondoc.backend.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -56,6 +58,83 @@ public class FriendServiceImpl implements FriendService{
                 .build();
     }
 
+    @Override
+    @Transactional
+    public FriendRequestDto.Response requestDelete(Long userId, Long requestId) {
+        Friend friend = friendRepository.findByIdAndStatus(requestId, 0)
+                .orElseThrow(() -> new NotFoundException("요청이 없습니다."));
+
+        if(friend.getUser().getId() != userId){
+            throw new NoSuchElementException("요청에 대한 접근 권한이 없습니다.");
+        }
+
+        // 삭제 수행
+        friendRepository.delete(friend);
+
+        return FriendRequestDto.Response.builder()
+                .success(true)
+                .msg("성공적으로 요청을 삭제했습니다.")
+                .build();
+    }
+
+    // 친구 삭제
+    @Override
+    @Transactional
+    public FriendRequestDto.Response friendDelete(Long userId, String id) {
+        Long ID = Long.parseLong(id);
+        Friend friend = friendRepository.findById(ID)
+                .orElseThrow(() -> new NotFoundException("친구를 찾을 수 없습니다."));
+
+        if(friend.getFriendId() != userId && friend.getUser().getId() != userId){
+            throw new NoSuchElementException("삭제에 실패 했습니다.");
+        }
+
+        // 삭제 수행
+        friendRepository.delete(friend);
+
+        return FriendRequestDto.Response.builder()
+                .msg("삭제를 완료하였습니다.")
+                .success(true)
+                .build();
+    }
+
+    // 친구 검색
+    @Override
+    public FriendSearchDto.Response searchFriend(Long userId, String phoneNumber) {
+        // 나 검색
+        User me = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("유저를 찾을 수 없습니다."));
+
+        // 핸드폰 번호 기반으로 유저 탐색
+        User friend = userRepository.findByPhoneNumber(phoneNumber)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 회원입니다."));
+
+        // 양방향 검색 해야함
+        Friend result;
+        try{
+            // 친구 ID 기준으로 검색
+            result = friendRepository.findByFriendIdAndUserId(friend.getId(), userId)
+                    .orElseThrow(() -> new NotFoundException("친구를 찾을 수 없습니다."));
+
+        }catch(Exception e){
+            result = friendRepository.findByFriendIdAndUserId(userId, friend.getId())
+                    .orElseThrow(() -> new NotFoundException("친구를 찾을 수 없습니다."));
+        }
+
+        FriendSearchDto.FriendInfo friendInfo = FriendSearchDto.FriendInfo.builder()
+                .id(friend.getId())
+                .name(friend.getName())
+                .imageNumber(friend.getImageNumber() + "")
+                .phoneNumber(friend.getPhoneNumber())
+                .build();
+
+        return FriendSearchDto.Response.builder()
+                .success(true)
+                .msg("친구 검색에 성공하였습니다.")
+                .friend(friendInfo)
+                .build();
+    }
+
     // 요청 승인
     @Override
     @Transactional
@@ -89,23 +168,31 @@ public class FriendServiceImpl implements FriendService{
             throw new NoSuchElementException("요청에 접근 권한이 없습니다.");
         }
 
-        friendRepository.delete(friend);
+        if(friend.getStatus() == 0){
+            friendRepository.delete(friend);
+
+            return FriendRequestDto.Response.builder()
+                    .msg("거절이 완료 되었습니다.")
+                    .success(true)
+                    .build();
+        }
 
         return FriendRequestDto.Response.builder()
-                .msg("거절이 완료되었습니다.")
-                .success(true)
+                .msg("요청을 찾을 수 없습니다.")
+                .success(false)
                 .build();
+
     }
 
     // 받은 요청 목록
     @Override
-    public FriendListDto.Response receiveList(Long userId) {
+    public FriendRequestDto.RequestListResponse receiveList(Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("유저를 찾을 수 없습니다."));
 
         List<Friend> list = friendRepository.findByFriendIdAndStatus(userId, 0);
 
-        return FriendListDto.Response.builder()
+        return FriendRequestDto.RequestListResponse.builder()
                 .list(list)
                 .msg("친구 요청 목록을 불러왔습니다.")
                 .success(true)
@@ -114,13 +201,13 @@ public class FriendServiceImpl implements FriendService{
 
     // 보낸 요청 목록
     @Override
-    public FriendListDto.Response requestList(Long userId) {
+    public FriendRequestDto.RequestListResponse sendList(Long userId) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("유저를 찾을 수 없습니다."));
 
         List<Friend> list = friendRepository.findByUserAndStatus(user, 0);
 
-        return FriendListDto.Response.builder()
+        return FriendRequestDto.RequestListResponse.builder()
                 .list(list)
                 .msg("친구 요청 목록을 불러왔습니다.")
                 .success(true)
@@ -133,10 +220,34 @@ public class FriendServiceImpl implements FriendService{
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("유저를 찾을 수 없습니다."));
 
-        List<Friend> list = friendRepository.findByUserAndStatus(user, 1);
+        // 조회한 친구 목록 객체
+        List<Friend> list = friendRepository.findByUserOrFriendIdAndStatus(user, userId, 1);
+
+        // 응답을 위한 친구 목록 객체
+        List<FriendListDto.FriendInfoDto> friendList = new ArrayList<>();
+
+        for(Friend friend : list){
+            FriendListDto.FriendInfoDto temp;
+            // 현재 이 정보가 나의 정보
+            if(friend.getUser().getId().equals(userId)){
+                temp = FriendListDto.FriendInfoDto.builder()
+                        .id(friend.getId())
+                        .friendId(friend.getFriendId())
+                        .createdAt(friend.getCreatedAt())
+                        .build();
+            }else{
+                // 현재 이 정보가 상대의 정보
+                temp = FriendListDto.FriendInfoDto.builder()
+                        .id(friend.getId())
+                        .friendId(friend.getUser().getId())
+                        .createdAt(friend.getCreatedAt())
+                        .build();
+            }
+            friendList.add(temp);
+        }
 
         return FriendListDto.Response.builder()
-                .list(list)
+                .list(friendList)
                 .msg("친구 목록을 불러왔습니다.")
                 .success(true)
                 .build();
